ml-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
scikit-learn==1.3.2
pandas==2.1.2
numpy==1.26.1
joblib==1.3.2
pydantic==2.4.2
python-multipart==0.0.6
Pillow==10.4.0
